+++ src/main/kotlin/com/range/jarvana/controller/JarInfoController.kt
@@ -13,12 +13,15 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/info")
 @RestController
 class JarInfoController(private val service: JarMetadataService) {
-private val log = LoggerFactory.getLogger(javaClass)
+    private val log = LoggerFactory.getLogger(javaClass)
 
-@GetMapping("/all")
+    @GetMapping("/all")
     fun getAll(): ResponseEntity<ResponseDto<List<JarMetadataDto>>> {
+
         log.info("Getting all jars")
+
         val data = service.findAll()
+
         return ResponseEntity.ok(
             ResponseDto(
                 success = true,
@@ -30,8 +33,11 @@ private val log = LoggerFactory.getLogger(javaClass)
 
     @GetMapping("/{id}")
     fun getById(@PathVariable("id") id:Long):ResponseEntity<ResponseDto<JarMetadataDto>>{
+
         log.info("Getting artifact by id $id")
+
         val data = service.findById(id)
+
         return ResponseEntity.ok(ResponseDto(
             success = true,
             message = "Artifact by id $id",