+++ src/main/kotlin/com/range/jarvana/dto/JarDownloadDto.kt
@@ -1,0 +1,8 @@
+package com.range.jarvana.dto
+
+data class JarDownloadDto (
+    val filename: String,
+
+
+    val content: ByteArray
+)