+++ src/main/kotlin/com/range/jarvana/mapper/JarMapper.kt
@@ -1,5 +1,6 @@
 package com.range.jarvana.mapper
 
+import com.range.jarvana.dto.JarDownloadDto
 import com.range.jarvana.dto.JarMetadataDto
 import com.range.jarvana.model.JarFile
 
@@ -11,6 +12,13 @@ object JarMapper {
             name = jarFile.name,
             createdAt = jarFile.createdAt,
             sizeInKb = jarFile.size,
+        )
+    }
+    fun jarFileToJarDownload(jarFile: JarFile): JarDownloadDto {
+        return JarDownloadDto(
+            filename = jarFile.name,
+
+            content = jarFile.data
         )
     }
 }