+++ src/main/kotlin/com/range/jarvana/service/JarStorageService.kt
@@ -1,10 +1,11 @@
+import com.range.jarvana.dto.JarDownloadDto
 import com.range.jarvana.dto.JarMetadataDto
 import com.range.jarvana.model.JarFile
 import org.springframework.web.multipart.MultipartFile
 
 interface JarStorageService {
     fun upload(file: MultipartFile): JarMetadataDto
-    fun download(id: Long): JarFile
+    fun download(id: Long): JarDownloadDto
     fun delete(id: Long): JarMetadataDto
     fun deleteAll()
 }